refactor: extract a mixin class
add a mixin class to meld in models which generate forecast lines
