+++ project_forecast_line/models/hr_leave.py
@@ -8,7 +8,8 @@ _logger = logging.getLogger(__name__)
 
 
 class HrLeave(models.Model):
-    _inherit = "hr.leave"
+    _name = "hr.leave"
+    _inherit = ["hr.leave", "forecast.line.mixin"]
 
     @api.model_create_multi
     def create(self, vals_list):
@@ -80,17 +81,6 @@ class HrLeave(models.Model):
                 ]
             )
             to_update._update_forecast_lines()
-
-    def unlink(self):
-        ForecastLine = self.env["forecast.line"].sudo()
-        to_clean = ForecastLine.search(
-            [
-                ("res_model", "=", self._name),
-                ("res_id", "in", tuple(self.ids)),
-            ]
-        )
-        to_clean.unlink()
-        return super().unlink()
 
 
 # XXX: leave request should create forcast negative forecast?
