+++ project_forecast_line/models/project_project_stage.py
@@ -10,12 +10,3 @@ class ProjectProjectStage(models.Model):
         [("forecast", "Forecast"), ("confirmed", "Confirmed")],
         help="type of forecast lines created by the tasks of projects in that stage",
     )
-
-    # def write(self, values):
-    #     res = super().write(values)
-    #     if "forecast_line_type" in values:
-    #         projects = self.env["project.project"].search(
-    #             [("stage_id", "in", self.ids)]
-    #         )
-    #         projects.mapped("task_ids")._update_forecast_lines()
-    #     return res
